refactor(ai-context): 重构AI上下文工具目录结构并清理冗余文件
- 重构: 将AI上下文相关工具统一迁移到 scripts/ai/ 目录
- 清理: 删除冗余的CLAUDE文档文件和过时的图谱数据文件
- 整合: 合并AI上下文CLI工具为统一的session-tools
- 优化: 更新package.json的AI会话工具脚本配置
- 规范: 统一AI开发工具的目录结构和命名约定
