+++ package.json
@@ -19,12 +19,12 @@
     "url": "https://github.com/linch-tech/linch-kit.git"
   },
   "scripts": {
-    "linch": "node packages/core/dist/cli.js",
+    "linch": "bun packages/core/dist/cli.js",
     "build": "turbo run build",
     "build:packages": "turbo run build:packages --filter='./packages/*'",
     "build:apps": "turbo run build --filter='./apps/*'",
     "build:watch": "turbo run build:packages --filter='./packages/*' --watch",
-    "dev": "node scripts/dev-tools.js dev",
+    "dev": "bun scripts/dev-tools.js dev",
     "dev:apps": "turbo run dev --parallel",
     "dev:packages": "turbo run build:packages --filter='./packages/*' --watch",
     "lint": "turbo run lint",
@@ -39,21 +39,22 @@
     "clean:dist": "turbo run clean:dist",
     "clean:node_modules": "find . -name 'node_modules' -type d -exec rm -rf {} + 2>/dev/null || true",
     "reset": "bun run clean:dist && bun run clean:node_modules && bun install",
-    "deps:graph": "node scripts/deps-graph.js",
-    "deps:check": "node scripts/dev-tools.js check",
+    "deps:graph": "bun scripts/dev/deps-graph.js",
+    "deps:check": "bun scripts/dev-tools.js check",
     "changeset": "changeset",
     "changeset:version": "changeset version",
     "changeset:publish": "changeset publish",
-    "release": "node scripts/release.js",
-    "release:dry": "node scripts/release.js --dry-run",
+    "release": "bun scripts/release.js",
+    "release:dry": "bun scripts/release.js --dry-run",
     "version": "changeset version && bun install --lockfile-only",
     "ci": "bun install --frozen-lockfile && bun run build && bun run test && bun run lint",
     "prepare": "husky install",
-    "setup": "node scripts/update-all-configs.js && bun install",
-    "test:workflow": "node scripts/test-workflow.js",
+    "setup": "bun scripts/update-all-configs.js && bun install",
+    "test:workflow": "bun scripts/test-workflow.js",
     "validate": "bun run test:workflow",
-    "docs:cleanup": "node scripts/cleanup-docs.js",
-    "docs:migrate": "node scripts/migrate-docs.js",
+    "ai:session": "bun scripts/ai/session-tools.js",
+    "docs:cleanup": "bun scripts/cleanup-docs.js",
+    "docs:migrate": "bun scripts/migrate-docs.js",
     "docs:build": "turbo run build --filter='./apps/docs'",
     "docs:dev": "turbo run dev --filter='./apps/docs'"
   },
