--- packages/ai/NEXT_SESSION_PROMPT.md
@@ -1,237 +1,0 @@
-# LinchKit AI - Claude Code 上下文查询工具 - CLI实现完成
-
-**会话时间**: 2025-07-07  
-**当前分支**: `feature/vibe-coding-engine`  
-**项目状态**: Claude Code CLI集成已完成，可立即使用
-
----
-
-## 🎯 重大突破：CLI工具已可用！
-
-### ✅ 核心成果
-1. **独立CLI工具完成**: `ai-context-cli.js` 已创建并测试通过
-2. **Claude Code集成就绪**: JSON输出格式完全适合AI Agent调用
-3. **功能验证成功**: 能够查询并返回结构化数据
-4. **数据连接正常**: Neo4j连接稳定，5,446节点/7,969关系可访问
-
-### 🚀 Claude Code 使用方式
-
-**立即可用的命令**:
-```bash
-# 基础上下文查询
-bun ai-context-cli.js --query "用户认证" --limit 5
-
-# 代码模式查询
-bun ai-context-cli.js --query "React组件" --type patterns
-
-# 最佳实践查询  
-bun ai-context-cli.js --query "错误处理" --type practices
-
-# 人类友好的文本输出
-bun ai-context-cli.js --query "API设计" --format text
-```
-
-**JSON输出示例**:
-```json
-{
-  "success": true,
-  "query": "用户认证",
-  "type": "context", 
-  "timestamp": "2025-07-07T02:27:17.808Z",
-  "execution_time_ms": 301,
-  "data": {
-    "entities": [...],
-    "relationships": [...],
-    "metadata": {
-      "total_results": 20,
-      "relevance_score": 0.3
-    }
-  }
-}
-```
-
----
-
-## 🎉 与Gemini协商结果：CLI方案获胜
-
-基于Gemini的专业分析，我们选择了**直接CLI调用**而非MCP服务器：
-
-### ✅ CLI方案优势
-- **极简架构**: 零额外依赖，直接使用现有@linch-kit/ai包
-- **本地优先**: 完全在用户机器上运行，无需服务器进程
-- **易于调试**: Claude Code可以直接看到命令输出和错误
-- **即插即用**: 无需复杂的服务器配置和端口管理
-
-### 📋 架构决策
-- **放弃MCP服务器**: 避免过度工程化
-- **专注核心价值**: 让工具立即可用，而不是架构完美
-- **渐进式优化**: 先让Claude Code用起来，再根据反馈优化
-
----
-
-## ⚠️ 当前已知限制
-
-### 数据质量问题 (非阻塞)
-- **症状**: 实体名称显示为"Unknown"，但类型和ID信息正确
-- **影响**: 不影响Claude Code的使用，仍可通过type、package等获得有用信息
-- **计划**: 后续优化数据提取逻辑
-
-### 关系数据稀缺
-- **症状**: 查询结果中relationships为空数组
-- **影响**: 限制了代码关系分析，但实体查询正常工作
-- **计划**: 优化关系提取和查询算法
-
-### 中文意图识别
-- **症状**: 中文查询置信度较低(30%)
-- **影响**: 不影响结果返回，只是相关性评分偏低
-- **计划**: 改进中文语义理解算法
-
----
-
-## 🎯 下一阶段开发优先级
-
-### 优先级1: Claude Code使用体验优化
-- **目标**: 确保Claude Code能够顺利调用和解析结果
-- **任务**: 根据Claude Code实际使用反馈调整输出格式
-
-### 优先级2: 数据质量提升  
-- **目标**: 修复实体名称显示问题，提供更丰富的上下文
-- **任务**: 优化数据提取器，确保节点属性正确填充
-
-### 优先级3: 查询准确性改进
-- **目标**: 提升中文查询的相关性和准确性
-- **任务**: 改进意图识别和相关性评分算法
-
----
-
-## 🛠️ 立即行动计划
-
-### Session 1: Claude Code集成测试
-```bash
-# 在Claude Code中测试以下命令
-cd /home/laofahai/workspace/linch-kit
-
-# 测试基础查询
-bun ai-context-cli.js --query "authentication system"
-
-# 测试中文查询  
-bun ai-context-cli.js --query "用户管理" --limit 5
-
-# 测试不同类型
-bun ai-context-cli.js --query "React patterns" --type patterns
-```
-
-### Session 2: 反馈收集和优化
-- 根据Claude Code使用体验调整输出格式
-- 优化错误处理和超时机制
-- 改进查询响应时间
-
-### Session 3: 数据质量修复
-- 调试节点名称提取问题
-- 优化关系数据查询
-- 提升查询结果的丰富度
-
----
-
-## 🔧 开发环境配置
-
-### 环境变量 (已配置)
-```bash
-# .env.local
-NEO4J_CONNECTION_URI=neo4j+s://d4a26556.databases.neo4j.io
-NEO4J_USERNAME=neo4j  
-NEO4J_PASSWORD=UbY-DQI7y3TjkOw548Nm99Qpw43_DvAa2F6o8HRZBkY
-NEO4J_DATABASE=neo4j
-```
-
-### 快速验证命令
-```bash
-# 构建AI包
-bun run build --filter @linch-kit/ai
-
-# 测试CLI工具
-bun ai-context-cli.js --query "test" --limit 3
-
-# 检查连接状态  
-bun simple-test.js
-```
-
----
-
-## 📚 核心文件索引
-
-### 新增重要文件
-- **独立CLI工具**: `ai-context-cli.js` - Claude Code专用CLI
-- **简化测试**: `simple-test.js` - 快速功能验证
-- **MCP设计文档**: `packages/context-server/DESIGN.md` - 备用架构方案
-
-### 核心实现文件
-- **上下文查询**: `packages/ai/src/context/context-query-tool.ts`
-- **智能查询**: `packages/ai/src/query/intelligent-query-engine.ts`  
-- **Neo4j服务**: `packages/ai/src/graph/neo4j-service.ts`
-- **CLI命令**: `packages/ai/src/cli/commands/context.ts`
-
----
-
-## 🎯 成功指标 (已达成)
-
-### ✅ 技术指标
-- [x] **CLI工具可用**: 能够接收查询并返回JSON结果
-- [x] **Neo4j连接**: 稳定连接到图数据库
-- [x] **查询响应**: 平均300ms响应时间  
-- [x] **数据规模**: 5,446节点/7,969关系可访问
-
-### ✅ 集成指标
-- [x] **JSON输出**: 结构化数据格式适合AI解析
-- [x] **错误处理**: 失败时返回标准错误格式
-- [x] **参数支持**: 支持query、type、limit、format等参数
-- [x] **帮助系统**: 内置--help选项
-
----
-
-## 🔥 立即价值实现
-
-### 对Claude Code的即时价值
-1. **项目理解**: 快速获取LinchKit项目的代码实体和结构
-2. **上下文感知**: 基于自然语言查询获得相关代码上下文  
-3. **模式识别**: 查找项目中的设计模式和代码示例
-4. **最佳实践**: 获取特定场景的推荐做法
-
-### 技术创新点
-1. **Graph RAG实现**: 结合知识图谱和检索的创新应用
-2. **本地化优化**: 支持中文查询，适合本土开发
-3. **最小化设计**: 专注查询功能，避免功能膨胀  
-4. **CLI优先**: 简单直接的集成方式
-
----
-
-## 🔧 下次Session启动命令
-
-```bash
-# 快速启动命令 (复制运行)
-cd /home/laofahai/workspace/linch-kit && \
-git checkout feature/vibe-coding-engine && \
-echo "=== Claude Code CLI 工具状态检查 ===" && \
-echo "当前分支: $(git branch --show-current)" && \
-echo "工作目录: $(pwd)" && \
-echo "测试CLI工具..." && \
-bun ai-context-cli.js --query "LinchKit架构" --limit 3 && \
-echo "=== CLI工具已就绪，可开始Claude Code集成 ==="
-```
-
-**下次会话开始语**:
-```
-Claude Code 上下文查询CLI工具已完成并可用！当前在 feature/vibe-coding-engine 分支。CLI工具 `ai-context-cli.js` 已通过测试，可以立即为Claude Code提供项目上下文查询服务。请开始集成测试或根据使用反馈进行优化。
-```
-
----
-
-## 🎊 项目里程碑
-
-**🎉 重大成就**: 从零到可用CLI工具，LinchKit AI上下文查询功能已具备企业级应用能力！
-
-**🚀 下一目标**: 与Claude Code深度集成，收集真实使用反馈，持续优化用户体验。
-
-**💪 技术储备**: Graph RAG架构、Neo4j知识图谱、自然语言查询 - 为AI-First开发奠定坚实基础！
-
-祝Claude Code集成顺利！ 🎯
