+++ scripts/README.md
@@ -1,32 +1,59 @@
 # LinchKit 项目脚本
 
-## 📋 可用脚本
+## 📋 脚本目录结构
 
-### 🤖 AI 上下文工具
-位置：`ai-context/`
+```
+scripts/
+├── README.md              # 本文档
+├── ai/                    # AI相关工具
+│   ├── session-tools.js   # 主要AI助手工具
+│   └── context-cli.js     # 上下文查询工具
+└── dev/                   # 开发工具
+    └── deps-graph.js      # 依赖关系分析
+```
 
-- **ai-context-cli.js** - AI上下文查询工具（优化版，统一接口）
+## 🤖 AI 工具
 
-详见：[AI Context 工具文档](./ai-context/README.md)
-
-### 🔧 开发工具
-
-#### graph-data-extractor.ts
-**用途**: Neo4j图谱数据提取和更新
-**运行**: `bun scripts/graph-data-extractor.ts`
+### session-tools.js
+**主要AI助手工具集合**
+**运行**: `bun run ai:session <command>`
 **功能**: 
-- 扫描项目代码结构
-- 提取节点和关系数据
-- 更新Neo4j知识图谱
-- 支持增量更新
+- Session初始化和环境检查
+- 项目上下文查询（实体、符号、模式）
+- Neo4j图谱数据同步
+- 代码质量验证（基础/完整）
+- 分支管理
 
-#### deps-graph.js  
-**用途**: 包依赖关系分析
-**运行**: `node scripts/deps-graph.js`
+### context-cli.js
+**AI上下文查询工具**
+**运行**: `bun scripts/ai/context-cli.js`
+**功能**:
+- 查询实体定义和相关文件
+- 查询符号定义（函数、类、接口）
+- 查询实现模式
+
+## 🔧 开发工具
+
+### deps-graph.js  
+**包依赖关系分析**
+**运行**: `bun run deps:graph`
 **功能**:
 - 分析monorepo包依赖关系
 - 生成依赖图谱
 - 检测循环依赖
+
+## 📚 使用示例
+
+```bash
+# AI Session 工具
+bun run ai:session init "新功能开发"
+bun run ai:session query "User"
+bun run ai:session check
+bun run ai:session validate
+
+# 开发工具
+bun run deps:graph
+```
 
 #### dev-tools.js
 **用途**: 开发工具集合
