--- test-coverage-analysis-report.md
@@ -1,251 +1,0 @@
-# LinchKit 包测试覆盖率分析报告
-
-**生成时间**: 2025-07-04
-**分析范围**: packages/ 和 modules/ 目录
-**项目**: LinchKit AI-First 全栈开发框架
-
-## 📊 总体统计
-
-### 测试文件概览
-- **总测试文件数**: 31 个
-- **总源代码文件数**: 173 个 (.ts/.tsx)
-- **整体测试覆盖率**: 17.9% (31/173)
-
-### 包级别统计
-
-| 包名 | 测试文件数 | 源代码文件数 | 覆盖率 | 状态 |
-|------|------------|--------------|--------|------|
-| @linch-kit/auth | 9 | 19 | 47.4% | 🟢 良好 |
-| @linch-kit/core | 3 | 51 | 5.9% | 🔴 严重不足 |
-| @linch-kit/crud | 8 | 24 | 33.3% | 🟡 需改进 |
-| @linch-kit/schema | 5 | 28 | 17.9% | 🟡 需改进 |
-| @linch-kit/trpc | 5 | 5 | 100%* | 🟢 优秀 |
-| @linch-kit/ui | 1 | 44 | 2.3% | 🔴 严重不足 |
-| @linch-kit/create-linch-kit | 0 | 2 | 0% | 🔴 无测试 |
-
-*注：@linch-kit/trpc 的100%覆盖率是基于文件数量，不代表代码行覆盖率
-
-## 🔍 详细分析
-
-### 1. @linch-kit/core (🔴 严重不足)
-**覆盖率**: 5.9% (3/51)
-
-**现有测试文件**:
-- `logger.test.ts` - 日志模块
-- `manager.test.ts` - 配置管理
-- `registry.test.ts` - 插件注册
-
-**缺少测试的关键模块**:
-- `audit/` - 审计系统 (6个文件未测试)
-  - `audit-manager.ts` - 审计管理器
-  - `data-masker.ts` - 数据脱敏
-  - `factory.ts` - 工厂模式
-  - `server-factory.ts` - 服务器工厂
-  - `stores/database-store.ts` - 数据库存储
-  - `stores/file-store.ts` - 文件存储
-- `observability/` - 可观测性 (4个文件未测试)
-  - `health.ts` - 健康检查
-  - `metrics.ts` - 指标收集 (有.skip文件)
-  - `tracing.ts` - 链路追踪 (有.skip文件)
-  - `metrics-client-safe.ts` - 客户端安全指标
-- `config/` - 配置管理 (6个文件未测试)
-  - `nextjs-provider.ts` - NextJS提供者
-  - `nextjs-provider-client.ts` - NextJS客户端提供者
-  - `tenant-manager.ts` - 租户管理
-  - `simple-tenant-manager.ts` - 简单租户管理
-  - `watcher.ts` - 配置监听器
-- `utils/` - 工具函数 (5个文件未测试)
-  - `fs.ts` - 文件系统操作
-  - `network.ts` - 网络工具
-  - `network-client.ts` - 网络客户端
-- `cli/` - 命令行工具 (9个文件未测试)
-
-### 2. @linch-kit/ui (🔴 严重不足)
-**覆盖率**: 2.3% (1/44)
-
-**现有测试文件**:
-- `index.test.ts` - 入口文件基础测试
-
-**缺少测试的关键模块**:
-- `components/ui/` - UI组件库 (27个组件未测试)
-  - `button.tsx` - 按钮组件
-  - `form.tsx` - 表单组件
-  - `input.tsx` - 输入框组件
-  - `dialog.tsx` - 对话框组件
-  - `table.tsx` - 表格组件
-  - `select.tsx` - 选择器组件
-  - 等其他21个UI组件
-- `forms/` - 表单模块 (3个文件未测试)
-  - `form-wrapper.tsx` - 表单包装器
-  - `schema-field-renderer.tsx` - Schema字段渲染器
-  - `schema-form.tsx` - Schema表单
-- `hooks/` - React Hooks (2个文件未测试)
-  - `use-mobile.ts` - 移动端检测
-  - `use-toast.ts` - 消息提示
-- `utils/` - 工具函数 (2个文件未测试)
-  - `cn.ts` - 样式类名工具
-
-### 3. @linch-kit/schema (🟡 需改进)
-**覆盖率**: 17.9% (5/28)
-
-**现有测试文件**:
-- `entity.test.ts` - 实体模块
-- `field.test.ts` - 字段模块
-- `index.test.ts` - 入口文件
-- `schema-builder.test.ts` - Schema构建器
-- `schema.test.ts` - Schema核心
-
-**缺少测试的关键模块**:
-- `generators/` - 代码生成器 (4个文件未测试)
-  - `prisma.ts` - Prisma生成器
-  - `typescript.ts` - TypeScript生成器
-  - `base.ts` - 基础生成器
-- `migration/` - 迁移系统 (2个文件未测试)
-  - `migrator.ts` - 迁移器
-- `validation/` - 验证系统 (2个文件未测试)
-  - `validator.ts` - 验证器
-- `cli/` - 命令行工具 (2个文件未测试)
-- `infrastructure/` - 基础设施 (4个文件未测试)
-
-### 4. @linch-kit/crud (🟡 需改进)
-**覆盖率**: 33.3% (8/24)
-
-**现有测试文件**:
-- `factory.test.ts` - 工厂模式
-- `crud-manager.test.ts` - CRUD管理器
-- `permission-checker.test.ts` - 权限检查器
-- `cache-manager.test.ts` - 缓存管理器
-- `validation-manager.test.ts` - 验证管理器
-- `base-query-builder.test.ts` - 基础查询构建器
-- `condition-builder.test.ts` - 条件构建器
-- `prisma-query-builder.test.ts` - Prisma查询构建器
-
-**缺少测试的关键模块**:
-- `core/` - 核心模块 (4个文件未测试)
-  - `manager.ts` - 管理器
-  - `query-builder.ts` - 查询构建器
-  - `query-executor.ts` - 查询执行器
-  - `query-optimizer.ts` - 查询优化器
-  - `query-validator.ts` - 查询验证器
-- `plugins/` - 插件系统 (5个文件未测试)
-  - `base-plugin.ts` - 基础插件
-  - `hook-manager.ts` - 钩子管理器
-  - `audit-log-plugin.ts` - 审计日志插件
-  - `user-activity-plugin.ts` - 用户活动插件
-- `trpc/` - tRPC集成 (2个文件未测试)
-- `cli/` - 命令行工具 (1个文件未测试)
-
-### 5. @linch-kit/auth (🟢 良好)
-**覆盖率**: 47.4% (9/19)
-
-**现有测试文件**:
-- `index.test.ts` - 入口文件
-- `permission.middleware.test.ts` - 权限中间件
-- `casl-engine.test.ts` - CASL引擎
-- `mfa.test.ts` - 多因素认证
-- `enterprise.test.ts` - 企业功能
-- `nextauth-adapter.test.ts` - NextAuth适配器
-- `router-factory.test.ts` - 路由工厂
-- `permission.service.test.ts` - 权限服务
-- `enhanced-permission-engine.test.ts` - 增强权限引擎
-
-**缺少测试的模块**:
-- `cli/` - 命令行工具 (2个文件未测试)
-- `components/` - React组件 (1个文件未测试)
-- `trpc/router.ts` - tRPC路由
-- `plugin.ts` - 插件定义
-- 其他基础设施文件
-
-### 6. @linch-kit/trpc (🟢 优秀)
-**覆盖率**: 100% (5/5)
-
-**现有测试文件**:
-- `integration.test.ts` - 集成测试
-- `auth.test.ts` - 认证路由
-- `server.test.ts` - 服务器
-- `commands.test.ts` - 命令行
-- `crud.test.ts` - CRUD路由
-
-**状态**: 测试覆盖完整
-
-### 7. @linch-kit/create-linch-kit (🔴 无测试)
-**覆盖率**: 0% (0/2)
-
-**缺少测试的模块**:
-- `create.ts` - 创建工具
-- `index.ts` - 入口文件
-
-## 🎯 优先级建议
-
-### 高优先级 (立即处理)
-1. **@linch-kit/core** - 核心基础设施，测试覆盖率仅5.9%
-   - 优先测试：`config/manager.ts`, `audit/audit-manager.ts`, `observability/health.ts`
-   - 启用跳过的测试：`metrics.test.ts.skip`, `tracing.test.ts.skip`
-
-2. **@linch-kit/ui** - UI组件库，测试覆盖率仅2.3%
-   - 优先测试：表单组件、核心UI组件、hooks
-   - 重点：`forms/schema-form.tsx`, `hooks/use-toast.ts`
-
-3. **@linch-kit/create-linch-kit** - 无测试覆盖
-   - 添加基础功能测试
-
-### 中优先级 (近期处理)
-1. **@linch-kit/schema** - 提升到60%+覆盖率
-   - 重点：代码生成器、迁移系统、验证系统
-
-2. **@linch-kit/crud** - 提升到60%+覆盖率
-   - 重点：插件系统、查询优化器
-
-### 低优先级 (长期维护)
-1. **@linch-kit/auth** - 已有良好覆盖率，补充剩余模块
-2. **@linch-kit/trpc** - 保持当前覆盖率
-
-## 📋 建议的测试实施计划
-
-### 第一阶段 (立即开始)
-- [ ] 为 @linch-kit/core 的配置管理模块添加测试
-- [ ] 为 @linch-kit/ui 的核心组件添加测试
-- [ ] 为 @linch-kit/create-linch-kit 添加基础测试
-- [ ] 启用被跳过的测试文件
-
-### 第二阶段 (1-2周内)
-- [ ] 完善 @linch-kit/schema 的生成器测试
-- [ ] 完善 @linch-kit/crud 的插件系统测试
-- [ ] 添加集成测试用例
-
-### 第三阶段 (长期维护)
-- [ ] 提升代码行覆盖率至80%+
-- [ ] 添加性能测试
-- [ ] 添加端到端测试
-
-## 🔧 质量改进建议
-
-1. **设置测试覆盖率目标**
-   - 核心包 (core, schema, auth, crud): 90%+
-   - 工具包 (ui, trpc): 80%+
-   - 辅助包 (create-linch-kit): 70%+
-
-2. **建立测试规范**
-   - 单元测试：每个公共函数/类必须有测试
-   - 集成测试：关键业务流程必须有测试
-   - 组件测试：UI组件必须有渲染和交互测试
-
-3. **CI/CD集成**
-   - 在构建流程中强制测试覆盖率检查
-   - 设置覆盖率下降阈值报警
-
-4. **文档同步**
-   - 测试用例作为API使用示例
-   - 保持测试文档与代码同步
-
-## 🚨 风险提示
-
-1. **@linch-kit/core** 作为基础设施包，测试覆盖率极低存在高风险
-2. **@linch-kit/ui** 组件库缺少测试，影响用户体验保障
-3. **跳过的测试文件** 可能隐藏重要功能缺陷
-4. **整体覆盖率偏低** 不符合企业级框架的质量标准
-
----
-
-**报告生成**: 基于 LinchKit 项目 worktrees/test-coverage-packages-modules 分支分析
-**下次更新**: 建议每周更新测试覆盖率报告
